+++ src/MemoryRouterProvider/MemoryRouterProvider.tsx
@@ -1,10 +1,14 @@
 import React, { FC, ReactNode, useMemo } from "react";
 
 import { useMemoryRouter, MemoryRouter, Url, default as singletonRouter } from "../index";
+import { default as asyncSingletonRouter } from "../async";
 import { MemoryRouterEventHandlers } from "../useMemoryRouter";
 import { MemoryRouterContext } from "../MemoryRouterContext";
 
-type AbstractedNextDependencies = Pick<typeof import("next/dist/shared/lib/router-context"), "RouterContext">;
+type AbstractedNextDependencies = Pick<
+  typeof import("next/dist/shared/lib/router-context.shared-runtime"),
+  "RouterContext"
+>;
 
 export type MemoryRouterProviderProps = {
   /**
@@ -25,7 +29,7 @@ export function factory(dependencies: AbstractedNextDependencies) {
         return new MemoryRouter(url, async);
       }
       // Normally we'll just use the singleton:
-      return singletonRouter;
+      return async ? asyncSingletonRouter : singletonRouter;
     }, [url, async]);
 
     const routerSnapshot = useMemoryRouter(memoryRouter, eventHandlers);