+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # next-router-mock
 
+## 0.9.13
+
+### Patch Changes
+
+- 5afd47b: Add support for parsing `search` parameter for URLs
+
+## 0.9.12
+
+### Patch Changes
+
+- d2e8fff: Enable MemoryRouterProvider to use the async singleton
+
+## 0.9.11
+
+### Patch Changes
+
+- 4603169: fix: update UrlObject type to match native url module
+
+## 0.9.10
+
+### Patch Changes
+
+- c8270b3: Updated `MemoryRouterProvider` to support Next 13.5
+
 ## 0.9.9
 
 ### Patch Changes